Move BottomModal to features folder & clean some code
